cleaned up raw svg file for b52
The B52 svg is ready to get some processing
into a Ruby script. The idea is to resd the
inkscape geneated file into nokogiri, then
extract the path for the outline and do some
math on the path, rewriting it to make it
symmetrical. The way to do this is to manually
place a midpoint at the nose and tail, translate
them to (0,0), reverse the original points to
the midpoint, write their negative value
as the path, then translate back. This would
result in a nicely symmatric silhouette.